backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
flask-cors==4.0.1
pymongo==4.6.3
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.17
openai==1.40.0
scikit-learn==1.5.1
bcrypt==4.1.3
pydantic==2.8.2
ollama
